Updated my Unity (finally lol) and enabled reticle
- Using the GVR reticle. I had to add collision boxes to UI elements so
the reticle distance wasn't behind them, while still being visible and
causing a migraine lol. It's still not perfect, especially when looking
back and forth at the edge of the displays and the background. This
seems to be because the reticle size doesn't change based on depth (as
our eyes/brain would expect). Might have to look at the GVR Reticle
Pointer script to see if radius can be adjust by distance (w/o the
interaction torus effect).