Quiz transitions and button colors
Added the states, the state machine, the transitions between states, the triggers, and the code for setting triggers for the transiton animations. Started adding the transition animations. Changed the colors of non-quiz-answer buttons to teal with highlight color white. Changed the highlight color of quiz-answer buttons to a light green (did not modify those buttons' color once selected).